Implement "weaving" of native frames into stack
This patch implements the merging procedure of native frames into the
resulting stack, which is hereby called _weaving_.

The algorithm boils down to identifying where the Ruby -> C switch
occured, and vice versa. When we find a Cfunc frame when digging down
the Ruby stack, we try to find a match in the native stack based on the
function address. This is done by peeking into the
rb_callable_method_entry_struct returned from rb_profile_thread_frames()
(which is not public, so I copied struct definitions from ruby/ruby),
making this implementation obviously unstable.

The opposite (finding C -> Ruby transitions) is rather easy; we just
mark rb_vm_exec() calls as the switch.