Capture native stack frames (PCs)
Capture native stack frames (the current PC to be more specific) when
the sample capture process runs.

This implementation uses libbacktrace as its backend, which is the same
as the Rust implementation.

There should be no problems calling backtrace_simple() in the signal
handler as it is marked as async-signal-safe. The first 2 frames will be
skipped (not included in the capture) since they will always be
capture_native_backtrace() and sigprof_handler().