Declare pf2_ser_sample.ruby_thread_id as uintptr_t
The actual type of 'pthread_t' is implementation defined in POSIX. On
macOS, pthread_t is 'struct _opaque_pthread_t *', which is not
assignable to size_t (the type of ruby_thread_id), causing a compile
error.

This is not the case in Linux, where pthread_t is 'unsigned long'.

Pf2 relies on some sort of Ruby representation of pthread_t to be
available. Here we assume that pthread_t fits in uintptr_t (64 bits),
and convert it to a Integer using ULL2NUM.