Make maximum stack depth to capture configurable
Pf2 had a hard-coded limit of 200 Ruby frames and 300 native frames,
which was a conservative value chosen to keep memory consumption low.
That was a too low limit for some applications, including large views in
Rails applications.

This patch makes max_depth and max_native_depth configurable, allowing
stack size up to 8192 frames. There is no particular reason for 8192, it
just seemed to be a large enough number.